pyproject.toml
[project]
name = "sds-ax"
version = "0.1.0"
description = "SDS-AX: Advanced CLI Coding Agent built on DeepAgents"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.4",
    "langgraph>=1.0",
    "langchain>=1.0",
    "langchain-core>=1.0",
    "langchain-anthropic>=1.0",
    "tavily-python>=0.5",
    "httpx>=0.27",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
]

[project.scripts]
sds-ax = "sds_ax.cli:main"
